backend/requirements.txt
anyio==3.6.2
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.0.5
djangorestframework==3.13.1
dnspython==2.2.1
fastapi==0.89.1
idna==3.4
psycopg2==2.9.3
pydantic==1.10.4
pymongo==4.3.3
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
python-dotenv==0.21.0
pytz==2022.1
requests==2.28.1
shiboken6==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.22.0
text2num==2.4.0
tqdm==4.62.3
typing_extensions==4.4.0
urllib3==1.26.13
